% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing_fn.R
\name{cv_postprocess}
\alias{cv_postprocess}
\title{Postprocess CV scan data}
\usage{
cv_postprocess(pre_workup, rhe)
}
\arguments{
\item{pre_workup}{dataframe of raw CV data}

\item{rhe}{user inputted value of reference electrode shift}
}
\value{
post-processed CV data with one column per scan (excluding scan 1)
}
\description{
Postprocess CV scan data
}
